backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
cryptography==42.0.5
PyJWT==2.8.0
python-dotenv==1.0.1
pydantic==2.7.1